revision: organize
